Split out App.py for #53
